graphql-api-generator/requirements.txt
GraphQL-core-next==1.0.4
PyYAML